Refactor `systimer` and `timer` modules into a common `timer` module
