+++ esp-hal/src/timer/mod.rs
@@ -1,0 +1,6 @@
+//! General-purpose timers.
+
+#[cfg(systimer)]
+pub mod systimer;
+#[cfg(any(timg0, timg1))]
+pub mod timg;
