+++ examples/src/bin/embassy_hello_world.rs
@@ -19,7 +19,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 
 #[embassy_executor::task]
