+++ examples/src/bin/embassy_i2c.rs
@@ -28,7 +28,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 use lis3dh_async::{Lis3dh, Range, SlaveAddr};
