+++ examples/src/bin/embassy_i2c_bmp180_calibration_data.rs
@@ -29,7 +29,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 
 #[main]
