+++ examples/src/bin/embassy_i2s_read.rs
@@ -30,7 +30,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 use esp_println::println;
