+++ examples/src/bin/embassy_multicore.rs
@@ -25,7 +25,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 use esp_println::println;
 use static_cell::make_static;
