+++ examples/src/bin/embassy_multiprio.rs
@@ -32,7 +32,7 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::TimerGroup,
+    timer::timg::TimerGroup,
 };
 use esp_println::println;
 use static_cell::make_static;
